+++ code/pet.store.spring.internal/src/main/java/pet/store/spring/internal/services/interfaces/PetsDaoServiceI.java
@@ -4,6 +4,6 @@ import pet.store.spring.internal.model.interfaces.PetLogicEntityI;
 
 public interface PetsDaoServiceI {
 	public void create(PetLogicEntityI pet) throws Exception;
-	public PetLogicEntityI read(int id) throws Exception;
-	public void delete(int id) throws Exception;
+	public PetLogicEntityI read(long id) throws Exception;
+	public void delete(long id) throws Exception;
 }
